chore: move setup methods to test utils for reuse
