+++ test-integration/test-ledger-restore/tests/04_flexi_counter.rs
@@ -2,7 +2,13 @@ use std::{path::Path, process::Child};
 
 use cleanass::assert_eq;
 use integration_test_tools::{
-    expect, loaded_accounts::LoadedAccounts, tmpdir::resolve_tmp_dir,
+    expect,
+    loaded_accounts::LoadedAccounts,
+    scenario_setup::{
+        confirm_tx_with_payer_ephem, fetch_counter_ephem,
+        init_and_delegate_counter_and_payer,
+    },
+    tmpdir::resolve_tmp_dir,
     validator::cleanup,
 };
 use program_flexi_counter::{
@@ -12,8 +18,7 @@ use program_flexi_counter::{
 use solana_sdk::{pubkey::Pubkey, signer::Signer};
 use test_kit::init_logger;
 use test_ledger_restore::{
-    confirm_tx_with_payer_ephem, fetch_counter_ephem,
-    init_and_delegate_counter_and_payer, setup_offline_validator,
+    setup_offline_validator,
     setup_validator_with_local_remote_and_resume_strategy,
     wait_for_ledger_persist, TMP_DIR_LEDGER,
 };
