+++ test-integration/test-ledger-restore/tests/08_commit_update.rs
@@ -2,7 +2,13 @@ use std::{path::Path, process::Child};
 
 use cleanass::assert_eq;
 use integration_test_tools::{
-    expect, loaded_accounts::LoadedAccounts, tmpdir::resolve_tmp_dir,
+    expect,
+    loaded_accounts::LoadedAccounts,
+    scenario_setup::{
+        confirm_tx_with_payer_chain, confirm_tx_with_payer_ephem,
+        fetch_counter_chain, fetch_counter_ephem,
+    },
+    tmpdir::resolve_tmp_dir,
     validator::cleanup,
 };
 use program_flexi_counter::{
@@ -16,11 +22,9 @@ use solana_sdk::{
     native_token::LAMPORTS_PER_SOL, signature::Keypair, signer::Signer,
 };
 use test_ledger_restore::{
-    assert_counter_commits_on_chain, confirm_tx_with_payer_chain,
-    confirm_tx_with_payer_ephem, fetch_counter_chain, fetch_counter_ephem,
-    get_programs_with_flexi_counter, setup_validator_with_local_remote,
-    wait_for_cloned_accounts_hydration, wait_for_ledger_persist,
-    TMP_DIR_LEDGER,
+    assert_counter_commits_on_chain, get_programs_with_flexi_counter,
+    setup_validator_with_local_remote, wait_for_cloned_accounts_hydration,
+    wait_for_ledger_persist, TMP_DIR_LEDGER,
 };
 const COUNTER: &str = "Counter of Payer";
 fn payer_keypair() -> Keypair {
