+++ test-integration/test-ledger-restore/tests/13_timestamps_match_during_replay.rs
@@ -2,7 +2,10 @@ use std::{path::Path, process::Child};
 
 use cleanass::assert_eq;
 use integration_test_tools::{
-    expect, loaded_accounts::LoadedAccounts, tmpdir::resolve_tmp_dir,
+    expect,
+    loaded_accounts::LoadedAccounts,
+    scenario_setup::{airdrop_and_delegate_accounts, transfer_lamports},
+    tmpdir::resolve_tmp_dir,
     validator::cleanup,
 };
 use solana_sdk::{
@@ -12,8 +15,7 @@ use solana_sdk::{
 use solana_transaction_status::UiTransactionEncoding;
 use test_kit::init_logger;
 use test_ledger_restore::{
-    airdrop_and_delegate_accounts, setup_offline_validator,
-    setup_validator_with_local_remote, transfer_lamports,
+    setup_offline_validator, setup_validator_with_local_remote,
     wait_for_ledger_persist, SNAPSHOT_FREQUENCY, TMP_DIR_LEDGER,
 };
 use tracing::*;
