+++ test-integration/test-tools/src/lib.rs
@@ -3,13 +3,15 @@ pub mod dlp_interface;
 mod integration_test_context;
 pub mod loaded_accounts;
 mod run_test;
+pub mod toml_to_args;
+pub mod validator;
+pub mod scenario_setup;
 pub mod scheduled_commits;
 pub mod tmpdir;
 pub mod transactions;
 pub mod workspace_paths;
 
-pub mod toml_to_args;
-pub mod validator;
 pub use color_backtrace;
 pub use integration_test_context::IntegrationTestContext;
 pub use run_test::*;
+pub use validator::cleanup;
